requirements.txt
numpy
http://people.csail.mit.edu/hubert/pyaudio/packages/pyaudio-0.2.8.tar.gz
scikits.samplerate
requests>=1.2.0
google-cloud-speech==0.28.0
